Proyecto/pruebas proyecto/requirements.txt
phe==1.5.0
psycopg2-binary==2.9.11